+++ tests/testthat/test-add_resource.R
@@ -6,17 +6,16 @@ test_that("add_resource() returns a valid Data Package", {
   expect_true(check_package(add_resource(p, "new", df)))
   expect_true(check_package(add_resource(p, "new", df, schema)))
   expect_true(check_package(add_resource(p, "new", df_csv)))
+  expect_true(check_package(
+    add_resource(p, "new", df, title = "New", foo = "bar")
+  ))
 })
 
-test_that("add_resource() returns error on incorrect Data Package", {
+test_that("add_resource() returns error on invalid Data Package", {
   df <- data.frame("col_1" = c(1, 2), "col_2" = c("a", "b"))
   expect_error(
     add_resource(list(), "new", df),
-    paste(
-      "`package` must be a list describing a Data Package,",
-      "created with `read_package()` or `create_package()`."
-    ),
-    fixed = TRUE
+    class = "frictionless_error_package_invalid"
   )
 })
 
@@ -24,20 +23,47 @@ test_that("add_resource() returns error when resource name contains invalid
            characters", {
   p <- example_package
   df <- data.frame("col_1" = c(1, 2), "col_2" = c("a", "b"))
+
+  # Invalid names
   expect_error(
     add_resource(p, "New", df),
-    paste(
-      "`New` must only contain lowercase alphanumeric characters plus",
-      "`.`, `-` and `_`."
+    class = "frictionless_error_resource_name_invalid"
+  )
+  expect_error(
+    add_resource(p, "New", df),
+    regexp = paste(
+      "`resource_name` must only consist of lowercase alphanumeric characters,",
+      "\".\", \"-\" and \"_\"."
     ),
     fixed = TRUE
   )
-  expect_error(add_resource(p, "nëw", df), "only contain lowercase")
-  expect_error(add_resource(p, " new", df), "only contain lowercase")
-  expect_error(add_resource(p, "new ", df), "only contain lowercase")
-  expect_error(add_resource(p, "n ew", df), "only contain lowercase")
-  expect_error(add_resource(p, "n/ew", df), "only contain lowercase")
+  expect_error(
+    add_resource(p, "New", df),
+    regexp = "\"New\" does not meet those criteria.",
+    fixed = TRUE
+  )
+  expect_error(
+    add_resource(p, "nëw", df),
+    class = "frictionless_error_resource_name_invalid"
+  )
+  expect_error(
+    add_resource(p, " new", df),
+    class = "frictionless_error_resource_name_invalid"
+  )
+  expect_error(
+    add_resource(p, "new ", df),
+    class = "frictionless_error_resource_name_invalid"
+  )
+  expect_error(
+    add_resource(p, "n ew", df),
+    class = "frictionless_error_resource_name_invalid"
+  )
+  expect_error(
+    add_resource(p, "n/ew", df),
+    class = "frictionless_error_resource_name_invalid"
+  )
 
+  # Valid names
   expect_true(check_package(add_resource(p, "n.ew", df)))
   expect_true(check_package(add_resource(p, "n-ew", df)))
   expect_true(check_package(add_resource(p, "n_ew", df)))
@@ -51,7 +77,11 @@ test_that("add_resource() returns error when resource of that name already
   df <- data.frame("col_1" = c(1, 2), "col_2" = c("a", "b"))
   expect_error(
     add_resource(p, "deployments", df),
-    "`package` already contains a resource named `deployments`.",
+    class = "frictionless_error_resource_already_exists"
+  )
+  expect_error(
+    add_resource(p, "deployments", df),
+    regexp = "`package` already contains a resource named \"deployments\".",
     fixed = TRUE
   )
 })
@@ -61,7 +91,11 @@ test_that("add_resource() returns error when data is not data frame or
   p <- example_package
   expect_error(
     add_resource(p, "new", list()),
-    "`data` must be a data frame or path(s) to CSV file(s).",
+    class = "frictionless_error_data_type_invalid"
+  )
+  expect_error(
+    add_resource(p, "new", list()),
+    regexp = "`data` must either be a data frame or path(s) to CSV file(s).",
     fixed = TRUE
   )
 })
@@ -72,52 +106,44 @@ test_that("add_resource() returns error on invalid or empty data frame", {
   schema <- create_schema(df)
   expect_error(
     add_resource(p, "new", data.frame("col_1" = character(0))),
-    "`data` must be a data frame containing data.",
-    fixed = TRUE
+    class = "frictionless_error_data_invalid"
   )
   expect_error(
     add_resource(p, "new", data.frame("col_1" = character(0)), schema),
-    "`data` must be a data frame containing data.",
-    fixed = TRUE
+    class = "frictionless_error_data_invalid"
   )
 
   # For more tests see test-check_schema.R
 })
 
 test_that("add_resource() returns error if CSV file cannot be found", {
-  testthat::skip_if_offline()
+  skip_if_offline()
   p <- example_package
   df_csv <- test_path("data/df.csv")
   schema <- create_schema(data.frame("col_1" = c(1, 2), "col_2" = c("a", "b")))
   expect_error(
     add_resource(p, "new", "no_such_file.csv"),
-    "Can't find file at `no_such_file.csv`.",
-    fixed = TRUE
+    class = "frictionless_error_path_not_found"
   )
   expect_error(
     add_resource(p, "new", "no_such_file.csv", schema),
-    "Can't find file at `no_such_file.csv`.",
-    fixed = TRUE
+    class = "frictionless_error_path_not_found"
   )
   expect_error(
     add_resource(p, "new", c(df_csv, "no_such_file.csv")),
-    "Can't find file at `no_such_file.csv`.",
-    fixed = TRUE
+    class = "frictionless_error_path_not_found"
   )
   expect_error(
     add_resource(p, "new", c("no_such_file.csv", df_csv)),
-    "Can't find file at `no_such_file.csv`.",
-    fixed = TRUE
+    class = "frictionless_error_path_not_found"
   )
   expect_error(
     add_resource(p, "new", c("no_such_file_1.csv", "no_such_file_2.csv")),
-    "Can't find file at `no_such_file_1.csv`.",
-    fixed = TRUE
+    class = "frictionless_error_path_not_found"
   )
   expect_error(
     add_resource(p, "new", "http://example.com/no_such_file.csv"),
-    "Can't find file at `http://example.com/no_such_file.csv`.",
-    fixed = TRUE
+    class = "frictionless_error_url_not_found"
   )
 })
 
@@ -131,28 +157,65 @@ test_that("add_resource() returns error on mismatching schema and data", {
   # df
   expect_error(
     add_resource(p, "new", df, schema_invalid),
-    paste(
-      "Field names in `schema` must match column names in data:",
-      "ℹ Field names: `no_such_col`, `col_2`",
-      "ℹ Column names: `col_1`, `col_2`",
-      sep = "\n"
-    ),
-    fixed = TRUE
+    class = "frictionless_error_fields_colnames_mismatch"
   )
 
   # csv
   expect_error(
     add_resource(p, "new", df_csv, schema_invalid),
-    paste(
-      "Field names in `schema` must match column names in data:",
-      "ℹ Field names: `no_such_col`, `col_2`",
-      "ℹ Column names: `col_1`, `col_2`",
-      sep = "\n"
-    ),
-    fixed = TRUE
+    class = "frictionless_error_fields_colnames_mismatch"
   )
 
   # For more tests see test-check_schema.R
+})
+
+test_that("add_resource() returns error if ... arguments are unnamed", {
+  p <- create_package()
+  df <- data.frame("col_1" = c(1, 2), "col_2" = c("a", "b"))
+  schema <- create_schema(df)
+  expect_error(
+    add_resource(p, "new", df, schema, delim = ",", "unnamed_value"),
+    class = "frictionless_error_unnamed_argument"
+  )
+  expect_error(
+    add_resource(p, "new", df, schema, delim = ",", "unnamed_value"),
+    "All arguments in `...` must be named.",
+    fixed = TRUE
+  )
+})
+
+test_that("add_resource() returns error if ... arguments are reserved", {
+  p <- create_package()
+  df <- data.frame("col_1" = c(1, 2), "col_2" = c("a", "b"))
+  expect_error(
+    add_resource(p, "new", df, name = "custom_name"),
+    class = "frictionless_error_resource_properties_reserved"
+  )
+  expect_error(
+    add_resource(p, "new", df, name = "custom_name", format = "custom_format"),
+    class = "frictionless_error_resource_properties_reserved"
+  )
+
+  expect_error(
+    add_resource(p, "new", df, name = "custom_name"),
+    regexp = "`name` must be removed as argument.",
+    fixed = TRUE
+  )
+  expect_error(
+    add_resource(p, "new", df, name = "custom_name"),
+    regexp = "name is automatically added as resource property.",
+    fixed = TRUE
+  )
+  expect_error(
+    add_resource(p, "new", df, name = "custom_name", format = "custom_format"),
+    regexp = "`name` and `format` must be removed as arguments.",
+    fixed = TRUE
+  )
+  expect_error(
+    add_resource(p, "new", df, name = "custom_name", format = "custom_format"),
+    regexp = "name and format are automatically added as resource properties.",
+    fixed = TRUE
+  )
 })
 
 test_that("add_resource() adds resource", {
@@ -227,7 +290,7 @@ test_that("add_resource() can add resource from data frame, readable by
 
 test_that("add_resource() can add resource from local, relative, absolute,
            remote or compressed CSV file, readable by read_resource()", {
-  testthat::skip_if_offline()
+  skip_if_offline()
   p <- example_package
   schema <- get_schema(p, "deployments")
 
@@ -334,4 +397,20 @@ test_that("add_resource() sets correct properties for CSV resources", {
   expect_identical(p$resources[[6]]$mediatype, "text/tab-separated-values")
   expect_identical(p$resources[[6]]$encoding, "UTF-8")
   expect_identical(read_resource(p, "df_delim_2"), read_resource(p, "df"))
+})
+
+test_that("add_resource() sets ... arguments as extra properties", {
+  p <- create_package()
+  df <- data.frame("col_1" = c(1, 2), "col_2" = c("a", "b"))
+  df_csv <- test_path("data/df.csv")
+
+  # df
+  p <- add_resource(p, "new_df", df, title = "custom_title", foo = "bar")
+  expect_identical(p$resources[[1]]$title, "custom_title")
+  expect_identical(p$resources[[1]]$foo, "bar")
+
+  # csv
+  p <- add_resource(p, "new_csv", df_csv, title = "custom_title", foo = "bar")
+  expect_identical(p$resources[[2]]$title, "custom_title")
+  expect_identical(p$resources[[2]]$foo, "bar")
 })